DockerTest/Server/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
pandas==2.1.1
scikit-learn==1.3.1